Harden server-messages validation and live verification
Validate module imports before replacing existing copies and fail fast on
bad server-messages schedules. This keeps test and push workflows from
silently destroying modules or relying on unverified runtime behavior.
