+++ modules/server-messages/src/functions/server-message-helpers.js
@@ -1,24 +1,48 @@
 import { takaro } from '@takaro/helpers';
+import {
+  computeFingerprint,
+  createExecutionLockHeartbeat,
+  DEFAULT_INTERVAL,
+  DEFAULT_TIME_ZONE,
+  findUnknownPlaceholders,
+  getInitialState,
+  getIntervalStatus,
+  getNextSelection,
+  isValidTimeZone,
+  MAX_MESSAGES,
+  MAX_WEIGHT,
+  normalizeInterval,
+  normalizeMessages,
+  normalizeOrder,
+  normalizeTimeZone,
+  renderPlaceholders,
+  SUPPORTED_PLACEHOLDERS,
+} from './server-message-utils.js';
+
+export {
+  computeFingerprint,
+  createExecutionLockHeartbeat,
+  DEFAULT_INTERVAL,
+  DEFAULT_TIME_ZONE,
+  findUnknownPlaceholders,
+  getInitialState,
+  getIntervalStatus,
+  getNextSelection,
+  isValidTimeZone,
+  MAX_MESSAGES,
+  MAX_WEIGHT,
+  normalizeInterval,
+  normalizeMessages,
+  normalizeOrder,
+  normalizeTimeZone,
+  renderPlaceholders,
+  SUPPORTED_PLACEHOLDERS,
+};
 
 export const STATE_KEY = 'server_messages_state';
 export const FINGERPRINT_KEY = 'server_messages_fingerprint';
 export const LOCK_KEY = 'server_messages_lock';
-export const DEFAULT_INTERVAL = '*/15 * * * *';
-export const DEFAULT_TIME_ZONE = 'UTC';
-export const MAX_MESSAGES = 100;
-export const MAX_WEIGHT = 20;
-export const SUPPORTED_PLACEHOLDERS = ['playerCount', 'serverName'];
 const LOCK_TIMEOUT_MS = 15 * 60 * 1000;
-const HEARTBEAT_INTERVAL_MS = 60 * 1000;
-const WEEKDAY_TO_CRON_DAY = {
-  Sun: 0,
-  Mon: 1,
-  Tue: 2,
-  Wed: 3,
-  Thu: 4,
-  Fri: 5,
-  Sat: 6,
-};
 
 async function findVariable(gameServerId, moduleId, key) {
   const res = await takaro.variable.variableControllerSearch({
@@ -85,63 +109,7 @@ export async function refreshExecutionLock(gameServerId, moduleId, token) {
 }
 
 export function startExecutionLockHeartbeat(gameServerId, moduleId, token) {
-  let stopped = false;
-  let heartbeatInFlight = null;
-  let timerHandle = null;
-
-  function clearHeartbeatTimer() {
-    if (timerHandle && typeof clearTimeout === 'function') {
-      clearTimeout(timerHandle);
-    }
-    timerHandle = null;
-  }
-
-  function scheduleNextHeartbeat() {
-    if (stopped || typeof setTimeout !== 'function') return;
-
-    clearHeartbeatTimer();
-    timerHandle = setTimeout(() => {
-      heartbeat().catch((err) => {
-        console.error(`server-message-helpers: execution lock heartbeat failed: ${err}`);
-      });
-    }, HEARTBEAT_INTERVAL_MS);
-  }
-
-  async function heartbeat() {
-    if (stopped) return false;
-    if (heartbeatInFlight) return heartbeatInFlight;
-
-    heartbeatInFlight = (async () => {
-      const refreshed = await refreshExecutionLock(gameServerId, moduleId, token);
-      if (refreshed) {
-        scheduleNextHeartbeat();
-      } else {
-        clearHeartbeatTimer();
-      }
-      return refreshed;
-    })();
-
-    try {
-      return await heartbeatInFlight;
-    } finally {
-      heartbeatInFlight = null;
-    }
-  }
-
-  async function stopHeartbeat() {
-    stopped = true;
-    clearHeartbeatTimer();
-    if (heartbeatInFlight) {
-      await heartbeatInFlight.catch(() => false);
-    }
-  }
-
-  scheduleNextHeartbeat();
-
-  return {
-    heartbeat,
-    stopHeartbeat,
-  };
+  return createExecutionLockHeartbeat(() => refreshExecutionLock(gameServerId, moduleId, token));
 }
 
 export async function releaseExecutionLock(gameServerId, moduleId, token) {
@@ -173,359 +141,4 @@ export async function getState(gameServerId, moduleId) {
 
 export async function setState(gameServerId, moduleId, state) {
   await upsertVariable(gameServerId, moduleId, STATE_KEY, JSON.stringify(state));
-}
-
-export function normalizeMessages(messages) {
-  if (!Array.isArray(messages)) return [];
-
-  return messages
-    .filter((message) => message && typeof message.text === 'string')
-    .slice(0, MAX_MESSAGES)
-    .map((message) => ({
-      text: message.text,
-      weight: normalizeWeight(message.weight),
-    }));
-}
-
-export function normalizeOrder(order) {
-  return order === 'random' ? 'random' : 'sequential';
-}
-
-export function normalizeInterval(interval) {
-  if (typeof interval !== 'string') return DEFAULT_INTERVAL;
-  const trimmed = interval.trim();
-  return trimmed || DEFAULT_INTERVAL;
-}
-
-export function normalizeWeight(weight) {
-  const parsed = Number(weight);
-  if (!Number.isInteger(parsed) || parsed < 1) return 1;
-  return Math.min(parsed, MAX_WEIGHT);
-}
-
-export function normalizeTimeZone(timeZone) {
-  if (typeof timeZone !== 'string') return DEFAULT_TIME_ZONE;
-  const trimmed = timeZone.trim();
-  return trimmed || DEFAULT_TIME_ZONE;
-}
-
-export function isValidTimeZone(timeZone) {
-  try {
-    Intl.DateTimeFormat('en-US', {
-      timeZone,
-      minute: 'numeric',
-      hour: 'numeric',
-      day: 'numeric',
-      month: 'numeric',
-      weekday: 'short',
-      hourCycle: 'h23',
-    }).formatToParts(new Date());
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-export function computeFingerprint(order, messages) {
-  return hashString(JSON.stringify({ order, messages }));
-}
-
-export function hashString(input) {
-  let hash = 0x811c9dc5;
-
-  for (let i = 0; i < input.length; i++) {
-    hash ^= input.charCodeAt(i);
-    hash = Math.imul(hash, 0x01000193) >>> 0;
-  }
-
-  return hash.toString(16).padStart(8, '0');
-}
-
-export function getInitialState(order, messages) {
-  if (order === 'random') {
-    return {
-      order: 'random',
-      bag: shuffleBag(messages),
-      cursor: 0,
-    };
-  }
-
-  return {
-    order: 'sequential',
-    index: 0,
-  };
-}
-
-export function getNextSelection(order, messages, state) {
-  if (messages.length === 0) return null;
-
-  if (order === 'random') {
-    let workingState = state;
-    if (!isValidRandomState(workingState, messages.length, messages)) {
-      console.warn('server-message-helpers: random state bag is inconsistent with configured weights, rebuilding');
-      workingState = getInitialState(order, messages);
-    }
-
-    if (!workingState.bag.length || workingState.cursor >= workingState.bag.length) {
-      workingState = getInitialState(order, messages);
-    }
-
-    const selectedIndex = workingState.bag[workingState.cursor];
-    if (!Number.isInteger(selectedIndex) || selectedIndex < 0 || selectedIndex >= messages.length) {
-      workingState = getInitialState(order, messages);
-    }
-
-    return {
-      messageIndex: workingState.bag[workingState.cursor],
-      nextState: {
-        order: 'random',
-        bag: [...workingState.bag],
-        cursor: workingState.cursor + 1,
-      },
-      bagSize: workingState.bag.length,
-      cursor: workingState.cursor,
-    };
-  }
-
-  let index = Number.isInteger(state?.index) ? state.index : 0;
-  if (index < 0 || index >= messages.length) index = 0;
-
-  return {
-    messageIndex: index,
-    nextState: {
-      order: 'sequential',
-      index: (index + 1) % messages.length,
-    },
-  };
-}
-
-export function findUnknownPlaceholders(text) {
-  const unknown = new Set();
-
-  text.replace(/\{([a-zA-Z0-9_]+)\}/g, (_match, key) => {
-    if (!SUPPORTED_PLACEHOLDERS.includes(key)) {
-      unknown.add(key);
-    }
-    return _match;
-  });
-
-  return [...unknown].sort();
-}
-
-export function renderPlaceholders(text, context) {
-  return text.replace(/\{([a-zA-Z0-9_]+)\}/g, (match, key) => {
-    if (key === 'playerCount') return String(context.playerCount);
-    if (key === 'serverName') return context.serverName;
-    return match;
-  });
-}
-
-export function shuffleBag(messages) {
-  const bag = [];
-
-  messages.forEach((message, index) => {
-    for (let i = 0; i < message.weight; i++) {
-      bag.push(index);
-    }
-  });
-
-  for (let i = bag.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    const tmp = bag[i];
-    bag[i] = bag[j];
-    bag[j] = tmp;
-  }
-
-  return bag;
-}
-
-export function getIntervalStatus(interval, now = new Date(), timeZone = DEFAULT_TIME_ZONE) {
-  const normalized = normalizeInterval(interval);
-  const parts = normalized.split(/\s+/);
-  if (parts.length !== 5) {
-    return { valid: false, matches: false, normalized };
-  }
-
-  const timeParts = getCronTimeParts(now, timeZone);
-  if (!timeParts) {
-    return { valid: false, matches: false, normalized };
-  }
-
-  const [minuteField, hourField, dayOfMonthField, monthField, dayOfWeekField] = parts;
-  const minuteMatches = matchesCronField(minuteField, timeParts.minute, 0, 59);
-  const hourMatches = matchesCronField(hourField, timeParts.hour, 0, 23);
-  const monthMatches = matchesCronField(monthField, timeParts.month, 1, 12);
-  const dayOfMonthMatches = matchesCronField(dayOfMonthField, timeParts.dayOfMonth, 1, 31);
-  const dayOfWeekMatches = matchesCronField(dayOfWeekField, timeParts.dayOfWeek, 0, 7, { normalizeDayOfWeek: true });
-
-  if (
-    minuteMatches === null ||
-    hourMatches === null ||
-    monthMatches === null ||
-    dayOfMonthMatches === null ||
-    dayOfWeekMatches === null
-  ) {
-    return { valid: false, matches: false, normalized };
-  }
-
-  const daysMatch = getDayMatchStatus(dayOfMonthField, dayOfMonthMatches, dayOfWeekField, dayOfWeekMatches);
-
-  return {
-    valid: true,
-    matches: minuteMatches && hourMatches && monthMatches && daysMatch,
-    normalized,
-  };
-}
-
-function getCronTimeParts(now, timeZone) {
-  try {
-    const formatter = new Intl.DateTimeFormat('en-US', {
-      timeZone,
-      minute: 'numeric',
-      hour: 'numeric',
-      day: 'numeric',
-      month: 'numeric',
-      weekday: 'short',
-      hourCycle: 'h23',
-    });
-
-    const formattedParts = formatter.formatToParts(now);
-    const getPart = (type) => formattedParts.find((part) => part.type === type)?.value;
-    const weekday = getPart('weekday');
-
-    return {
-      minute: Number(getPart('minute')),
-      hour: Number(getPart('hour')),
-      dayOfMonth: Number(getPart('day')),
-      month: Number(getPart('month')),
-      dayOfWeek: WEEKDAY_TO_CRON_DAY[weekday] ?? Number.NaN,
-    };
-  } catch {
-    return null;
-  }
-}
-
-function getDayMatchStatus(dayOfMonthField, dayOfMonthMatches, dayOfWeekField, dayOfWeekMatches) {
-  const dayOfMonthIsWildcard = isWildcardField(dayOfMonthField);
-  const dayOfWeekIsWildcard = isWildcardField(dayOfWeekField);
-
-  if (dayOfMonthIsWildcard && dayOfWeekIsWildcard) return true;
-  if (dayOfMonthIsWildcard) return dayOfWeekMatches;
-  if (dayOfWeekIsWildcard) return dayOfMonthMatches;
-  return dayOfMonthMatches || dayOfWeekMatches;
-}
-
-function isWildcardField(field) {
-  return field.trim() === '*';
-}
-
-function matchesCronField(field, value, min, max, options = {}) {
-  const allowedValues = parseCronField(field, min, max, options);
-  if (!allowedValues) return null;
-  return allowedValues.has(value);
-}
-
-function parseCronField(field, min, max, options = {}) {
-  const segments = field.split(',').map((segment) => segment.trim());
-  if (segments.length === 0 || segments.some((segment) => !segment)) return null;
-
-  const allowedValues = new Set();
-
-  for (const segment of segments) {
-    const segmentValues = parseCronSegmentValues(segment, min, max, options);
-    if (!segmentValues) return null;
-
-    for (const entry of segmentValues) {
-      allowedValues.add(entry);
-    }
-  }
-
-  return allowedValues;
-}
-
-function parseCronSegmentValues(segment, min, max, options = {}) {
-  if (!segment) return null;
-
-  const stepParts = segment.split('/');
-  if (stepParts.length > 2) return null;
-
-  const [base, rawStep] = stepParts;
-  const step = rawStep === undefined ? 1 : parseCronNumberToken(rawStep);
-  if (!Number.isInteger(step) || step < 1 || step > max) return null;
-
-  if (base === '*') {
-    return buildCronValueSet(min, max, min, max, step, options);
-  }
-
-  if (base.includes('-')) {
-    const rangeParts = base.split('-');
-    if (rangeParts.length !== 2) return null;
-
-    const [startRaw, endRaw] = rangeParts;
-    const start = parseCronNumberToken(startRaw);
-    const end = parseCronNumberToken(endRaw);
-    if (!isAllowedCronNumber(start, min, max, options) || !isAllowedCronNumber(end, min, max, options)) {
-      return null;
-    }
-    if (start >= end) return null;
-
-    return buildCronValueSet(start, end, min, max, step, options);
-  }
-
-  if (rawStep !== undefined) return null;
-
-  const exact = parseCronNumberToken(base);
-  if (!isAllowedCronNumber(exact, min, max, options)) return null;
-  return new Set([normalizeCronValue(exact, options)]);
-}
-
-function parseCronNumberToken(token) {
-  if (!/^\d+$/.test(token)) return Number.NaN;
-  return Number(token);
-}
-
-function buildCronValueSet(start, end, min, max, step, options = {}) {
-  const values = new Set();
-
-  for (let current = start; current <= end; current += step) {
-    if (!isAllowedCronNumber(current, min, max, options)) return null;
-    values.add(normalizeCronValue(current, options));
-  }
-
-  return values;
-}
-
-function isAllowedCronNumber(value, min, max, options = {}) {
-  if (!Number.isInteger(value)) return false;
-  if (options.normalizeDayOfWeek && value === 7) return true;
-  return value >= min && value <= max;
-}
-
-function normalizeCronValue(value, options = {}) {
-  if (options.normalizeDayOfWeek && value === 7) return 0;
-  return value;
-}
-
-function isValidRandomState(state, messageCount, messages) {
-  if (!state || state.order !== 'random') return false;
-  if (!Array.isArray(state.bag)) return false;
-  if (!Number.isInteger(state.cursor) || state.cursor < 0) return false;
-  if (state.cursor > state.bag.length) return false;
-
-  const expectedCounts = new Array(messageCount).fill(0);
-  for (let index = 0; index < messages.length; index++) {
-    expectedCounts[index] = messages[index].weight;
-  }
-
-  const actualCounts = new Array(messageCount).fill(0);
-  for (const entry of state.bag) {
-    if (!Number.isInteger(entry) || entry < 0 || entry >= messageCount) return false;
-    actualCounts[entry] += 1;
-  }
-
-  if (state.bag.length !== expectedCounts.reduce((sum, count) => sum + count, 0)) {
-    return false;
-  }
-
-  return actualCounts.every((count, index) => count === expectedCounts[index]);
 }
